Strider mounting fixes; update mappings
This commit refactors health visual logic to make it a global system for each living entity.
