+++ connector/src/main/java/org/geysermc/connector/entity/LivingEntity.java
@@ -27,15 +27,23 @@ package org.geysermc.connector.entity;
 
 import com.github.steveice10.mc.protocol.data.game.entity.metadata.EntityMetadata;
 import com.nukkitx.math.vector.Vector3f;
+import com.nukkitx.protocol.bedrock.data.AttributeData;
 import com.nukkitx.protocol.bedrock.data.entity.EntityData;
 import com.nukkitx.protocol.bedrock.data.inventory.ContainerId;
 import com.nukkitx.protocol.bedrock.data.inventory.ItemData;
 import com.nukkitx.protocol.bedrock.packet.MobArmorEquipmentPacket;
 import com.nukkitx.protocol.bedrock.packet.MobEquipmentPacket;
+import com.nukkitx.protocol.bedrock.packet.UpdateAttributesPacket;
 import lombok.Getter;
 import lombok.Setter;
+import org.geysermc.connector.entity.attribute.AttributeType;
 import org.geysermc.connector.entity.type.EntityType;
 import org.geysermc.connector.network.session.GeyserSession;
+import org.geysermc.connector.utils.AttributeUtils;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 @Getter
 @Setter
@@ -97,5 +105,35 @@ public class LivingEntity extends Entity {
         session.sendUpstreamPacket(armorEquipmentPacket);
         session.sendUpstreamPacket(handPacket);
         session.sendUpstreamPacket(offHandPacket);
+    }
+
+    @Override
+    public void updateBedrockAttributes(GeyserSession session) {
+        if (!valid) return;
+
+        float maxHealth = this.attributes.containsKey(AttributeType.MAX_HEALTH) ? this.attributes.get(AttributeType.MAX_HEALTH).getValue() : getDefaultMaxHealth();
+
+        List<AttributeData> attributes = new ArrayList<>();
+        for (Map.Entry<AttributeType, org.geysermc.connector.entity.attribute.Attribute> entry : this.attributes.entrySet()) {
+            if (!entry.getValue().getType().isBedrockAttribute())
+                continue;
+
+            attributes.add(AttributeUtils.getBedrockAttribute(entry.getValue()));
+        }
+        // Add health attribute to properly show hearts when mounting
+        attributes.add(new AttributeData("minecraft:health", 0.0f, maxHealth, metadata.getFloat(EntityData.HEALTH, 20f), maxHealth));
+
+        UpdateAttributesPacket updateAttributesPacket = new UpdateAttributesPacket();
+        updateAttributesPacket.setRuntimeEntityId(geyserId);
+        updateAttributesPacket.setAttributes(attributes);
+        session.sendUpstreamPacket(updateAttributesPacket);
+    }
+
+    /**
+     * Used for the health visual when mounting an entity.
+     * @return the default maximum health for the entity.
+     */
+    protected float getDefaultMaxHealth() {
+        return 20f;
     }
 }
