+++ connector/src/main/java/org/geysermc/connector/entity/living/animal/horse/AbstractHorseEntity.java
@@ -27,23 +27,12 @@ package org.geysermc.connector.entity.living.animal.horse;
 
 import com.github.steveice10.mc.protocol.data.game.entity.metadata.EntityMetadata;
 import com.nukkitx.math.vector.Vector3f;
-import com.nukkitx.protocol.bedrock.data.AttributeData;
 import com.nukkitx.protocol.bedrock.data.entity.EntityFlag;
-import com.nukkitx.protocol.bedrock.packet.UpdateAttributesPacket;
-import org.geysermc.connector.entity.attribute.AttributeType;
 import org.geysermc.connector.entity.living.animal.AnimalEntity;
 import org.geysermc.connector.entity.type.EntityType;
 import org.geysermc.connector.network.session.GeyserSession;
-import org.geysermc.connector.utils.AttributeUtils;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 public class AbstractHorseEntity extends AnimalEntity {
-
-    // For updating the horse visual easier
-    private float health = 20f;
 
     public AbstractHorseEntity(long entityId, long geyserId, EntityType entityType, Vector3f position, Vector3f motion, Vector3f rotation) {
         super(entityId, geyserId, entityType, position, motion, rotation);
@@ -51,11 +40,6 @@ public class AbstractHorseEntity extends AnimalEntity {
 
     @Override
     public void updateBedrockMetadata(EntityMetadata entityMetadata, GeyserSession session) {
-
-        if (entityMetadata.getId() == 8) {
-            health = (float) entityMetadata.getValue();
-            updateBedrockAttributes(session);
-        }
 
         if (entityMetadata.getId() == 16) {
             byte xd = (byte) entityMetadata.getValue();
@@ -70,26 +54,5 @@ public class AbstractHorseEntity extends AnimalEntity {
         metadata.getFlags().setFlag(EntityFlag.WASD_CONTROLLED, true);
 
         super.updateBedrockMetadata(entityMetadata, session);
-    }
-
-    @Override
-    public void updateBedrockAttributes(GeyserSession session) {
-        if (!valid) return;
-
-        float maxHealth = attributes.containsKey(AttributeType.MAX_HEALTH) ? attributes.get(AttributeType.MAX_HEALTH).getValue() : 20f;
-
-        List<AttributeData> attributesLocal = new ArrayList<>();
-        for (Map.Entry<AttributeType, org.geysermc.connector.entity.attribute.Attribute> entry : this.attributes.entrySet()) {
-            if (!entry.getValue().getType().isBedrockAttribute())
-                continue;
-
-            attributesLocal.add(AttributeUtils.getBedrockAttribute(entry.getValue()));
-        }
-        attributesLocal.add(new AttributeData("minecraft:health", 0.0f, maxHealth, health, maxHealth));
-
-        UpdateAttributesPacket updateAttributesPacket = new UpdateAttributesPacket();
-        updateAttributesPacket.setRuntimeEntityId(geyserId);
-        updateAttributesPacket.setAttributes(attributesLocal);
-        session.sendUpstreamPacket(updateAttributesPacket);
     }
 }
