+++ connector/src/main/java/org/geysermc/connector/network/translators/java/entity/JavaEntitySetPassengersTranslator.java
@@ -140,8 +140,14 @@ public class JavaEntitySetPassengersTranslator extends PacketTranslator<ServerEn
             case ARMOR_STAND:
                 yOffset = 1.3f;
                 break;
+            case STRIDER:
+                yOffset = 2.8200102f;
+                break;
         }
         Vector3f offset = Vector3f.from(0f, yOffset, 0f);
+        if (mountType == EntityType.STRIDER) {
+            offset = offset.add(0f, 0f, -0.2f);
+        }
         // Without the X offset, more than one entity on a boat is stacked on top of each other
         if (rider && moreThanOneEntity) {
             offset = offset.add(Vector3f.from(0.2, 0, 0));
